move libsql vector store to core
