+++ packages/core/src/vector/libsql/README.md
