+++ packages/core/src/vector/libsql/filter.test.ts
