+++ packages/core/src/vector/libsql/filter.ts
