+++ packages/core/src/vector/libsql/index.test.ts
