+++ packages/core/src/vector/libsql/index.ts
@@ -1,6 +1,7 @@
 import { createClient, type Client as TursoClient, type InValue } from '@libsql/client';
-import { Filter } from '@mastra/core/filter';
-import { MastraVector, type IndexStats, type QueryResult } from '@mastra/core/vector';
+
+import { Filter } from '../../filter';
+import { MastraVector, type IndexStats, type QueryResult } from '../index';
 
 import { LibSQLFilterTranslator } from './filter';
 import { buildFilterQuery } from './sql-builder';
