+++ packages/core/src/vector/libsql/sql-builder.ts
