--- vector-stores/libsql/.gitignore
@@ -1,1 +1,0 @@
-.env
