--- vector-stores/libsql/CHANGELOG.md
@@ -1,204 +1,0 @@
-# @mastra/vector-libsql
-
-## 0.1.0-alpha.25
-
-### Patch Changes
-
-- Updated dependencies [8151f44]
-- Updated dependencies [e897f1c]
-- Updated dependencies [3700be1]
-  - @mastra/core@0.2.0-alpha.90
-
-## 0.1.0-alpha.24
-
-### Patch Changes
-
-- Updated dependencies [27275c9]
-  - @mastra/core@0.2.0-alpha.89
-
-## 0.1.0-alpha.23
-
-### Patch Changes
-
-- ccbc581: Updated operator validation and handling for all vector stores
-- Updated dependencies [ccbc581]
-  - @mastra/core@0.2.0-alpha.88
-
-## 0.1.0-alpha.22
-
-### Patch Changes
-
-- Updated dependencies [7365b6c]
-  - @mastra/core@0.2.0-alpha.87
-
-## 0.1.0-alpha.21
-
-### Minor Changes
-
-- 5916f9d: Update deps from fixed to ^
-
-### Patch Changes
-
-- 7f24c29: Add Chroma Filter translator and updated vector store tests
-- Updated dependencies [6fa4bd2]
-- Updated dependencies [e2e76de]
-- Updated dependencies [7f24c29]
-- Updated dependencies [67637ba]
-- Updated dependencies [04f3171]
-  - @mastra/core@0.2.0-alpha.86
-
-## 0.0.1-alpha.20
-
-### Patch Changes
-
-- Updated dependencies [e9d1b47]
-  - @mastra/core@0.2.0-alpha.85
-
-## 0.0.1-alpha.19
-
-### Patch Changes
-
-- Updated dependencies [2f17a5f]
-- Updated dependencies [cb290ee]
-- Updated dependencies [b4d7416]
-- Updated dependencies [38b7f66]
-  - @mastra/core@0.2.0-alpha.84
-
-## 0.0.1-alpha.18
-
-### Patch Changes
-
-- 34048bb: Hook up Astra filter and added tests
-- 78eec7c: Started implementation on Unified Filter API for several vector stores.
-- 72f7fb9: Updated LibSQL filter store to add back in $not
-- 9625602: Use mastra core splitted bundles in other packages
-- Updated dependencies [30322ce]
-- Updated dependencies [78eec7c]
-- Updated dependencies [9625602]
-- Updated dependencies [8769a62]
-  - @mastra/core@0.2.0-alpha.83
-
-## 0.0.1-alpha.17
-
-### Patch Changes
-
-- Updated dependencies [73d112c]
-  - @mastra/core@0.1.27-alpha.82
-
-## 0.0.1-alpha.16
-
-### Patch Changes
-
-- Updated dependencies [9fb3039]
-  - @mastra/core@0.1.27-alpha.81
-
-## 0.0.1-alpha.15
-
-### Patch Changes
-
-- b422ed3: Bundle vector provider packages with tsup
-
-## 0.0.1-alpha.14
-
-### Patch Changes
-
-- Updated dependencies [327ece7]
-  - @mastra/core@0.1.27-alpha.80
-
-## 0.0.1-alpha.13
-
-### Patch Changes
-
-- Updated dependencies [21fe536]
-  - @mastra/core@0.1.27-alpha.79
-
-## 0.0.1-alpha.12
-
-### Patch Changes
-
-- Updated dependencies [685108a]
-- Updated dependencies [685108a]
-  - @mastra/core@0.1.27-alpha.78
-
-## 0.0.1-alpha.11
-
-### Patch Changes
-
-- Updated dependencies [8105fae]
-  - @mastra/core@0.1.27-alpha.77
-
-## 0.0.1-alpha.10
-
-### Patch Changes
-
-- 7a469e7: Bump vectors
-
-## 0.0.1-alpha.9
-
-### Patch Changes
-
-- Updated dependencies [ae7bf94]
-- Updated dependencies [ae7bf94]
-  - @mastra/core@0.1.27-alpha.76
-
-## 0.0.1-alpha.8
-
-### Patch Changes
-
-- Updated dependencies [23dcb23]
-  - @mastra/core@0.1.27-alpha.75
-
-## 0.0.1-alpha.7
-
-### Patch Changes
-
-- Updated dependencies [7b87567]
-  - @mastra/core@0.1.27-alpha.74
-
-## 0.0.1-alpha.6
-
-### Patch Changes
-
-- Updated dependencies [3427b95]
-  - @mastra/core@0.1.27-alpha.73
-
-## 0.0.1-alpha.5
-
-### Patch Changes
-
-- Updated dependencies [e4d4ede]
-- Updated dependencies [06b2c0a]
-  - @mastra/core@0.1.27-alpha.72
-
-## 0.0.1-alpha.4
-
-### Patch Changes
-
-- Updated dependencies [d9c8dd0]
-  - @mastra/core@0.1.27-alpha.71
-
-## 0.0.1-alpha.3
-
-### Patch Changes
-
-- 3c9a389: Added more connection options
-
-## 0.0.1-alpha.2
-
-### Patch Changes
-
-- bdaf834: publish packages
-
-## 0.0.1-alpha.1
-
-### Patch Changes
-
-- Updated dependencies [dd6d87f]
-- Updated dependencies [04434b6]
-  - @mastra/core@0.1.27-alpha.70
-
-## 0.0.1-alpha.0
-
-### Patch Changes
-
-- 0d5a03d: Vector store modules
