--- vector-stores/libsql/package.json
@@ -1,32 +1,0 @@
-{
-  "name": "@mastra/vector-libsql",
-  "version": "0.1.0-alpha.25",
-  "description": "LibSQL vector store provider for Mastra",
-  "type": "module",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/index.d.ts",
-        "default": "./dist/index.js"
-      }
-    },
-    "./package.json": "./package.json"
-  },
-  "scripts": {
-    "build": "tsup-node src/index.ts --format esm --dts --clean --treeshake",
-    "dev": "tsup-node src/index.ts --format esm --dts --clean --treeshake --watch",
-    "test": "vitest run"
-  },
-  "dependencies": {
-    "@libsql/client": "^0.14.0",
-    "@mastra/core": "workspace:^"
-  },
-  "devDependencies": {
-    "@tsconfig/recommended": "^1.0.7",
-    "@types/node": "^22.9.0",
-    "tsup": "^8.0.1",
-    "vitest": "^3.0.4"
-  }
-}
